Simulation

One function handles running the simulation: simulate.

SystemsOfSystems.simulate — Function

simulate(user_data; t, init_fcn, rates_fcn, updates_fcn, close_fcn, seed, options)

Runs a simulation, returning the time history, end time, and final model.

user_data: Can be anything used by the init_fcn

t: A collection of monotonic times. The sim will step to exactly each given time, plus as many other steps are required by the solver and models. At the very least, this must contain a start time and end time.

init_fcn: Will be called with (t_start, user_data, seed), where t_start is the first element of the above t input. This must return a ModelDescription.

rates_fcn: Will be called with (t, model) and is expected to return a RatesOutput.

updates_fcn: Will be called with (t, model) and is expected to return an UpdatesOutput, or nothing when there are no updates, outputs, replacement t_next, or stop request.

close_fcn: Will be called when simulation completes (even if an error is caught) with (t, model). No return value is expected.

seed: A top-level seed (Int) to control all random number generation in the sim. The init_fcn receives this as a BranchingSeed.

options: See SimOptions.